+++ src_new/aliases/levi_ataxia/for_levi/levi_062424/157_Scytherus_(All_Classes).lua
@@ -1,0 +1,17 @@
+--[[mudlet
+type: alias
+name: Scytherus (All Classes)
+hierarchy:
+- Levi_Ataxia
+attributes:
+  isActive: 'yes'
+  isFolder: 'no'
+regex: ^srr$
+command: ''
+packageName: ''
+]]--
+
+if gmcp.Char.Status.class == "Serpent" then
+  serp_setmode_scytherus()
+  serp_ekanelia_offense()
+end